fix : build errors
